fix(invariants): restore memory and SDK boundaries
