+++ .gitignore
@@ -7,6 +7,7 @@ firebase-debug.*.log
 functions/firebase-debug.log
 functions/firebase-debug.*.log
 .next
+.next-launch
 .next *
 .next.__trash__*
 out
