+++ next.config.js
@@ -1,5 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
+  distDir: '.next-launch',
   output: 'export',
   trailingSlash: true,
   images: {
