+++ src/lib/auth.ts
@@ -1,43 +1,8 @@
-import { useState, useEffect } from 'react';
-import { 
-  signInWithEmailAndPassword, 
-  createUserWithEmailAndPassword,
-  signOut,
-  onAuthStateChanged,
-  User
-} from 'firebase/auth';
-import { auth } from './firebase';
+export {
+  onAuthChanged,
+  signInWithEmail as signIn,
+  signOut as logout,
+  signUpWithEmail as signUp,
+} from '@/services/auth/authService';
 
-export const signIn = (email: string, password: string) => {
-  return signInWithEmailAndPassword(auth, email, password);
-};
-
-export const signUp = (email: string, password: string) => {
-  return createUserWithEmailAndPassword(auth, email, password);
-};
-
-export const logout = () => {
-  return signOut(auth);
-};
-
-export const onAuthChanged = (callback: (user: User | null) => void) => {
-  return onAuthStateChanged(auth, callback);
-};
-
-export const useAuth = () => {
-  const [user, setUser] = useState<User | null>(null);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, (user) => {
-      setUser(user);
-      setLoading(false);
-    });
-
-    return () => unsubscribe();
-  }, []);
-
-  return { user, loading };
-};
-
-export { auth };
+export { useAuth } from '@/hooks/useAuth';
