+++ src/types/khepera.ts
@@ -36,16 +36,9 @@ export type EntryAnchor = {
 };
 
 export interface KheperaMemory {
-  userId: string;
-  sessionCount: number;
   recurringThemes: ThemeTag[];
   recentThemes: ThemeTag[];
   dominantTone: EmotionalTone;
-  previousTone?: EmotionalTone;
-  recentStances: ResponseStance[];
-  recentStyles: KheperaStyleProfile[];
-  lastReturnType?: 'immediate' | 'delayed';
-  lastSessionDate: Date | null;
 }
 
 export interface TextStructureFeatures {
